rename: 색상 선택기 이름 변경
