+++ src/components/settings/SettingsColorThemeSelector.tsx
@@ -2,25 +2,20 @@ import React, { useEffect, useState } from 'react';
 import SettingsSingleSelect from './SettingsSingleSelect';
 import { useNavigation } from '@react-navigation/native';
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import {
-  getSelectedColorThemeSetting,
-  setSelectedColorThemeSetting,
-  type ColorThemeSetting,
-} from '@storage/settings';
+import { getSelectedColorThemeSetting, setSelectedColorThemeSetting, type ColorThemeSetting } from '@storage/settings';
 import { APP_COLOR_THEME_OPTIONS } from '@styles/appTheme';
 import type { RootStackParamList } from '@navigation/AppNavigator';
 import { useIapContext } from '@provider/IapProvider';
 
 /** 어쩜레드·애니블루만 무료 사용자에게 허용 (APP_COLOR_THEME_OPTIONS 상단 2개와 동일) */
 const FREE_COLOR_THEME_VALUES = new Set<ColorThemeSetting>(['awesomeRed', 'anyBlue']);
-
-const SettingsThemeSelector: React.FC = () => {
+/**
+ * 설정 화면 — 앱 색상 테마(프라이머리 팔레트) 선택 전용
+ */
+const SettingsColorThemeSelector: React.FC = () => {
   const { premiumUnlocked } = useIapContext();
-  const navigation =
-    useNavigation<NativeStackNavigationProp<RootStackParamList>>();
-  const [selectedTheme, setSelectedTheme] = useState<ColorThemeSetting>(() =>
-    getSelectedColorThemeSetting()
-  );
+  const navigation = useNavigation<NativeStackNavigationProp<RootStackParamList>>();
+  const [selectedTheme, setSelectedTheme] = useState<ColorThemeSetting>(() => getSelectedColorThemeSetting());
 
   useEffect(() => {
     const stored = getSelectedColorThemeSetting();
@@ -37,8 +32,7 @@ const SettingsThemeSelector: React.FC = () => {
     setSelectedColorThemeSetting(value);
   };
 
-  const isOptionLocked = (value: ColorThemeSetting) =>
-    !premiumUnlocked && !FREE_COLOR_THEME_VALUES.has(value);
+  const isOptionLocked = (value: ColorThemeSetting) => !premiumUnlocked && !FREE_COLOR_THEME_VALUES.has(value);
 
   return (
     <SettingsSingleSelect
@@ -52,4 +46,4 @@ const SettingsThemeSelector: React.FC = () => {
   );
 };
 
-export default SettingsThemeSelector;
+export default SettingsColorThemeSelector;
