+++ src/components/settings/index.ts
@@ -4,5 +4,5 @@ export { default as SettingsVoiceCommands } from './SettingsVoiceCommands';
 export { default as SettingsBackup } from './SettingsBackup';
 export { default as SettingsDataManagement } from './SettingsDataManagement';
 export { default as SettingsAppInfo } from './SettingsAppInfo';
-export { default as SettingsThemeSelector } from './SettingsThemeSelector';
+export { default as SettingsColorThemeSelector } from './SettingsColorThemeSelector';
 export { default as SettingsVersion } from './SettingsVersion';
